Systematise tag lookup
This commit brings a more systematic approach to detecting LN tags.
The names in the fields list now match those we feed as the final
PlainTextDocument's metadata, and we try to collect (but do not
necessarily use) all the fields it lists.

We also now store detected and potential tag matches as
attributes in the nodes set rather than as names to a text vector,
and handling of the slug is now consistent ("byline", not "BYLINE:"
or "byline:"), with a standardised lookup function.

Code (currently commented out) can detect potential tags which we
aren't currently catching: either because we haven't seen them
before or because the fields list doesn't yet have the translation.
This is a little noisy, especially with tabloid newspapers that
like headlines of the form "TRUMP: I AM THE GREATEST" (which looks
very much like a new LN "TRUMP" metadata tag) but is useful for
development.

The purpose of this reorganisation is to be more comprehensible,
more extensible in the future with new tag names or translations,
and to clear the decks to reason clearly about a future new main
content detection heuristic.