New main content detection
The previous approach to detecting potentially significant chunks
of the web page for extracting content from was to search for the
first <div> tag following each <br>. The main text was taken to
be the <div> with the longest content after the heading. This is
problematic for a few reasons:

Firstly, some of these <div>s were not quite empty enough to be
discarded, but too empty to have any content (such as
<div><br></div>). These were discarded by an earlier commit in
this series, as they sometimes masked content in the next <div>.

Secondly, and more importantly, not all LN documents have their
main content in a single top-level <div>. Articles with tabular
data, for example, may have several <table>s directly under the
<body> tag, which were not being read at all.

This commit changes the algorithm. Instead of looking at the
first <div> after each <br>, it now looks at all top-level non-<br>
elements after the first <br>; this catches all the <table>s etc.
Further, rather than assuming that there is a single top-level
<div> containing the main article body, it keeps track of each of
the elements parsed as something else (headline, date, other
metadata) and assumes that the actual text of the article is
whatever else is left once they are discarded.

I suspect we should limit only to the largest continuous sequence
of remaining content, in order to avoid accidentally collecting
new tags and other dross as part of the main content. Code is
currently in place to warn of non-contiguous sequences, to allow
for wide-scale testing.