Add random string to article IDs
It turns out that there is a non-hypothetical possibility of article
IDs to clash. As they're used to lookup from the Corpus downstream,
this is a significant problem as it essentially discards one of the
articles with clashing IDs.

Assume that you've collected The New York Times and The New York
Daily News. Both will reduce to "TheNewYo"; if you have collected
both papers on the same day, your IDs will be of the form
"TheNewYo20191213xxx" where xxx is simply the loop index from the
original file. This is not enough to ensure uniqueness, as it's
normally in the range 0-200.

This can also hit when inputting multiple articles from a *single*
large paper from the same day, when they split over two input files.

This commit adds a string of random letters to the end of the ID.