Change id algorithm to improve reproducibility from run-to-run
The algorithm for the corpus ids included 8 random letters to
greatly increase the probability that they were unique.
Unfortunately, this has the difficulty that if the corpus is
re-imported then the ids will change - which is somewhat
undesirable.

This commit changes the disambiguation letters to be a shortened
MD5 hash of the metadata, rather than 8 random characters. This
means that ids will only change if the content is actually
different.