Fix bugs with date parsing and with document ids
The date parsing code has been separated more cleanly between the python and native versions.

It's now also more robust to the python code misparsing: it has an unfortunate habit of picking up '26' or similar as the important part of a date when given ambiguous input, and treating that as 1900/01/01T02:26:00 or similar. We now reject any match which relies on less than 6 characters of source string, and fall back on our own code.

Finally, we fix a long-standing problem with the digests not matching up when (identical) documents have their IDs constructed. This was a result of bits of the underlying serialisation being in a different order despite the POSIXct values being identical. We now avoid this by doing as.character() first.