Better language handling
VCorpus feeds in a readerControl structure with language='en' by default, which is a major footgun if we rely on it in preference to whatever is in the data, when doing non-English work.

We now only use the declared language as a fallback if we can't get a language field from the document itself.