Better handling of language, where the Advance document doesn't provide one.
There are two sources of language info - the Nexis document itself, and the language code passed to our reader function by VCorpus(). The declared language tag can't be default, as tm defaults to passing 'en' in via VCorpus when nothing is given (which is a major potential footgun when dealing in non-English material). We previously ignored the value passed in, except when trying to work out how to parse the date and edition string. Now we use it, but only as a fallback if the document doesn't provide on (which it usually does).